Transferpak & Experimental Netplay Support (Currently disabled)

+++ GLideN64/src/MupenPlusPluginAPI.cpp
@@ -4,7 +4,8 @@
 #include "N64.h"
 #include <cstddef>
 
-extern uint32_t rdram_size;
+// Defined in mupen64plus-core main.c
+extern size_t rdram_size;
 
 extern "C" {
